Refactor RUCSS APIClient
