+++ inc/Engine/Optimization/RUCSS/Frontend/APIClient.php
@@ -1,23 +1,11 @@
 <?php
 declare(strict_types=1);
 
-namespace WP_Rocket\Engine\Optimization\RUCSS;
+namespace WP_Rocket\Engine\Optimization\RUCSS\Frontend;
 
-class APIClient {
-	/**
-	 * API URL.
-	 */
-	const API_URL = 'https://central-saas.wp-rocket.me/';
+use WP_Rocket\Engine\Optimization\RUCSS\AbstractAPIClient;
 
-	/**
-	 * SAAS main API path.
-	 */
-	const TREESHAKE_PATH = 'api';
-
-	/**
-	 * API Warmup path.
-	 */
-	const WARMUP_PATH = 'warmup';
+class APIClient extends AbstractAPIClient {
 
 	/**
 	 * Calls Central Saas API.
@@ -69,30 +57,4 @@ class APIClient {
 		];
 	}
 
-	/**
-	 * Handle Saas request error.
-	 *
-	 * @param  array|WP_Error $request WP Remote request.
-	 *
-	 * @return array
-	 */
-	protected function handle_request_error( $request ) {
-		if ( is_wp_error( $request ) ) {
-			return [
-				'code'    => 400,
-				'message' => $request->get_error_message(),
-			];
-		}
-
-		$code = wp_remote_retrieve_response_code( $request );
-
-		if ( 200 !== $code ) {
-			return [
-				'code'    => $code,
-				'message' => __( 'Remove Unused CSS is not available at the moment. Please retry later', 'rocket' ),
-			];
-		}
-
-		return false;
-	}
 }
