+++ inc/Engine/Optimization/RUCSS/ServiceProvider.php
@@ -24,14 +24,14 @@ class ServiceProvider extends AbstractServiceProvider {
 		'rucss_resources_table',
 		'rucss_database',
 		'rucss_admin_subscriber',
-		'rucss_api_client',
+		'rucss_frontend_api_client',
 		'rucss_used_css',
 		'rucss_used_css_query',
 		'rucss_frontend_subscriber',
 		'local_cache',
 		'rucss_resource_fetcher',
 		'rucss_resources_query',
-		'rucss_warmup_apiclient',
+		'rucss_warmup_api_client',
 	];
 
 	/**
@@ -51,7 +51,7 @@ class ServiceProvider extends AbstractServiceProvider {
 			->withArgument( $this->getContainer()->get( 'rucss_usedcss_table' ) );
 
 		$this->getContainer()->add( 'rucss_used_css_query', 'WP_Rocket\Engine\Optimization\RUCSS\Database\Query\UsedCSS' );
-		$this->getContainer()->add( 'rucss_api_client', 'WP_Rocket\Engine\Optimization\RUCSS\APIClient' );
+		$this->getContainer()->add( 'rucss_frontend_api_client', 'WP_Rocket\Engine\Optimization\RUCSS\Frontend\APIClient' );
 		$this->getContainer()->add( 'rucss_used_css_controller', 'WP_Rocket\Engine\Optimization\RUCSS\Controller\UsedCSS' )
 			->withArgument( $this->getContainer()->get( 'rucss_used_css_query' ) );
 
@@ -62,17 +62,17 @@ class ServiceProvider extends AbstractServiceProvider {
 		$this->getContainer()->share( 'rucss_frontend_subscriber', 'WP_Rocket\Engine\Optimization\RUCSS\Frontend\Subscriber' )
 			->withArgument( $this->getContainer()->get( 'options' ) )
 			->withArgument( $this->getContainer()->get( 'rucss_used_css_controller' ) )
-			->withArgument( $this->getContainer()->get( 'rucss_api_client' ) );
+			->withArgument( $this->getContainer()->get( 'rucss_frontend_api_client' ) );
 
 		$this->getContainer()->add( 'local_cache', '\WP_Rocket\Engine\Optimization\AssetsLocalCache' )
 			->withArgument( rocket_get_constant( 'WP_ROCKET_MINIFY_CACHE_PATH' ) )
 			->withArgument( rocket_direct_filesystem() );
 
-		$this->getContainer()->add( 'rucss_warmup_apiclient', '\WP_Rocket\Engine\Optimization\RUCSS\Warmup\APIClient' );
+		$this->getContainer()->add( 'rucss_warmup_api_client', '\WP_Rocket\Engine\Optimization\RUCSS\Warmup\APIClient' );
 
 		$this->getContainer()->add( 'rucss_resource_fetcher_process', '\WP_Rocket\Engine\Optimization\RUCSS\Warmup\ResourceFetcherProcess' )
 			->withArgument( $this->getContainer()->get( 'rucss_resources_query' ) )
-			->withArgument( $this->getContainer()->get( 'rucss_warmup_apiclient' ) );
+			->withArgument( $this->getContainer()->get( 'rucss_warmup_api_client' ) );
 
 		$this->getContainer()->share( 'rucss_resource_fetcher', '\WP_Rocket\Engine\Optimization\RUCSS\Warmup\ResourceFetcher' )
 			->withArgument( $this->getContainer()->get( 'local_cache' ) )
